feat(Admin Users Abandoned)Speed optimization

+++ app/Http/Controllers/Dashboard/AbandonedController.php
@@ -10,59 +10,59 @@ use App\Model\Event;
 use App\Model\Ticket;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Exports\AbandonedExport;
-
-
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Cache;
 
 class AbandonedController extends Controller
 {
     public function index(){
-        $data = [];
-        $list = ShoppingCart::with('user')->get();
 
+        // Before showing the abandoned cart, we need to check if the items inside the cart was a free event, or if it was finally booked by the user.
+        // We need to check since the last time we entered in this page. So, we can save this timestamp in the cache.
+        $timestamp = Cache::get('timestamp-last-time-check-abandoned-cart', Carbon::now()->subYears(10)->format('Y-m-d H:i:s'));
+        $list = ShoppingCart::with('user')->where('created_at', '>', $timestamp)->get();
+        $evids = [];
+        $data = [];
         $tickets = [];
         $ticks = Ticket::where('status', 1)->get();
-        $evids = [];
-
-        $freeEvents = Event::where('view_tpl','elearning_free')->pluck('id');
-        $freeEvents = $freeEvents->toArray();
+        if(count($list) > 0){
+            $lastShoppingCartChecked = $timestamp;
 
 
-        //dd($freeEvents);
-        //dd($list);
-        foreach ($list as $key => $item) {
-            $user_id = $item->identifier;
-            $cart = unserialize($item->content);
+            $freeEvents = Event::where('view_tpl','elearning_free')->pluck('id');
+            $freeEvents = $freeEvents->toArray();
 
-            $userEvents = isset($item->user[0]) ? $item->user[0]->events()->pluck('event_id')->toArray() : [];
+            foreach ($list as $key => $item) {
+                if($lastShoppingCartChecked < $item->created_at)
+                    $lastShoppingCartChecked = $item->created_at;
+                $user_id = $item->identifier;
+                $cart = unserialize($item->content);
 
+                $userEvents = isset($item->user[0]) ? $item->user[0]->events()->pluck('event_id')->toArray() : [];
 
-            foreach ($cart as $cartItem) {
+                foreach ($cart as $cartItem) {
 
-                if(in_array($cartItem->options['event'],$userEvents)){
-                    $item->delete();
-                    continue;
+                    if(in_array($cartItem->options['event'],$userEvents)){
+                        $item->delete();
+                        continue;
+                    }
+
+                    if(!in_array($cartItem->options['event'],$freeEvents)){
+                        $data['list'][$user_id] = $cartItem;
+                        $evids[] = $cartItem->options['event'];
+                    }
+
                 }
-
-                //dd(in_array($cartItem->options['event'],$freeEvents));
-                if(!in_array($cartItem->options['event'],$freeEvents)){
-                    //dd($cartItem->options['event']);
-                    $data['list'][$user_id] = $cartItem;
-                    //dd($cartItem->options['event']);
-                    $evids[] = $cartItem->options['event'];
-                }
-
             }
-
-
+            Cache::set('timestamp-last-time-check-abandoned-cart', $lastShoppingCartChecked);
         }
-        //dd($evids);
         $events = Event::whereIn('id', $evids)->get()->getDictionary();
         $data['events'] = $events;
         $data['tickets'] = $ticks->getDictionary();
-        $data['abcart'] = ShoppingCart::with('user')->get()->keyBy('identifier');
-        //dd($data['tickets']);
+        // $data['abcart'] = ShoppingCart::with('user')->get()->keyBy('identifier');
 
-        //dd($data);
+        unset($data['list']); // We don't want to show the previous table 
+        $data['list_by_shopping_cart'] = ShoppingCart::with('user')->get();
         return view('admin.abandoned.index', $data);
     }
